🎉 added example code; AcceptHeaderProcessor
We're starting a new project here, based on chleb-bible-search
We intend to be a fully-compatible project, but written entirely
in Java, because modern tooling doesn't support Perl and we can
take regular security fixes, and run on AWS Lambda and so on, if
we use Java.  It also increases the pool of available developers
we can call on.

This is fully experimental at this stage.  Please direct real requests
to the Perl related project:

https://github.com/daybologic/chleb-bible-search